+++ tools/automated-tests/package.json
@@ -15,6 +15,7 @@
         "fast-glob": "^3.3.3",
         "gray-matter": "^4.0.3",
         "js-yaml": "^4.1.0",
+        "minimist": "^1.2.8",
         "node-fetch": "^3.3.2",
         "puppeteer": "^23.10.2"
     }
